pyproject.toml
[tool.poetry]
name = "brms"
version = "0.1.0"
description = ""
authors = ["varunsinghal <varunsinghal15@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
black = "^23.3.0"
flake8 = "^6.0.0"
flask = "^2.3.2"
isort = "^5.12.0"
sqlalchemy = "^2.0.12"
click = "^8.1.3"
lxml = "^4.9.2"
factory-boy = "^3.2.1"
pytest = "^7.3.1"
pypika = "^0.48.9"
marshmallow = "^3.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
